Create wordiest_author
In this challenge from Khan Academy, a table was created for me. My first challenge was to select all the authors who have written more than 1 
million words, using GROUP BY and HAVING. My results table had to include the 'author' and their total word count as a 'total_words' column. The next
challenge was to select all the authors that write more than an average of 150,000 words per book. My results table had to include the 'author' and 
average words as an 'avg_words' column.