Create gradebook
/** In this challenge from Khan Academy, they created a database to track student grades, with their name, number grade, and what percent of activities
they've completed. For the first part I had to select all of the rows, and display the name, number_grade, and percent_completed, which I computed by 
multiplying and rounding the fraction_completed column. The goal of the second part is a table that shows how many students have earned which 
letter_grade. My output goal was the letter_grade by using CASE with the number_grade column, outputting 'A' for grades > 90, 'B' for grades > 80, 'C' 
for grades > 70, and 'F' otherwise. Then I used COUNT with GROUP BY to show the number of students with each of those grades. **/