Rename filtering_using_where to karaoke_song_selector
In this problem, Khan Academy created a table with songs, and I used queries to decide what songs to sing. For the first step, I selected all the song titles. Then I added another SELECT that uses OR to show the titles of the songs that have an 'epic' mood or a release date after 1990. Last, I added another SELECT that uses AND to show the titles of songs that are 'epic', and released after 1990, and less than 4 minutes long.