Create dynamic_documents
In this project Khan Academy created a database for a documents app, with rows for each document with it's title, content, and author. In this first
step, I used UPDATE to change the author to 'Jackie Draper' for all rows where it's currently 'Jackie Paper'. Then I re-selected all the rows to make
sure the table changed like I expected. Then I deleted a row, being very careful not to delete all the rows. I only deleted a row where the title is
'Things I'm Afraid Of'. Then I re-selected all the rows to make sure the table changed like I expected.